+++ src/test/TestGUI.java
@@ -1,4 +1,4 @@
-package process;
+package test;
 
 public class TestGUI {
     public static void main(String[] args) {
